Add first (semi-broken) version of desktop site
The desktop site is now generated by Pelican!

This version is _very_ broken, so for now it's just a proof of concept.
I hope to have the issues worked out by the end of the day so that the
desktop may share posts with the main site (and thus get promoted to the
top of the screen)