Finish desktop site!
Finish porting the desktop site. This includes images, theme
information, a new intro window, and a few other changes.

Remove the desktop and archives link from the bottom of the page. This
makes the navigation consistent (and focused on the top of the page
rather than the bottom _and_ the top).

Update windowTools/widgetTools to include a set of prefix variables.
This simplifies the installation of these tools, since it's not quite as
easy to just put them in pages/ as in the theme.

Update the blogtech post to reflect some of the changes over the past
few years.