Update desktop text, add pelican3, update blogtech
Update the text for the desktop intro window. It now says things.

Add the post on transitioning to Pelican, part 3

Update the old blogging tech post to clarify a few things. It was
written assuming that someone would read it in its original form as a
post on the first generation desktop site, rather than the more modern
hybrid (and later Pelican-based) site.