Add updated about page Hide desktop link on mobile
Add updated about page. This will replace the "about" screen on desktop.

Hide the "Desktop Site" button on screens less than 800 px wide. This
should stop it from showing up and wrecking the stylesheet on most
mobile devices. I do still need to fix the header so that adding links
doesn't break it on a mobile phone, but this is a start at any rate and
should be like this anyhow, whether it's updated for mobile or not.